+++ forge/src/main/java/dev/aika/smsn/forge/mixin/MixinInfo.java
@@ -6,8 +6,10 @@ import java.util.Map;
 
 public class MixinInfo {
     public static Map<String, String> getMixinClassNames() {
+        //noinspection SpellCheckingInspection
         return MapUtils.createMap(
                 "dev.aika.smsn.forge.mixin.ad_astra.StationLoaderMixin", "ad_astra",
+                "dev.aika.smsn.forge.mixin.additionalstructures.EventsMixin", "additionalstructures",
                 "dev.aika.smsn.forge.mixin.aetherteam.UserData$ServerMixin", "nitrogen",
                 "dev.aika.smsn.forge.mixin.alex.AlexsCavesWebHelperMixin", "alexscaves",
                 "dev.aika.smsn.forge.mixin.alex.CitadelConstantsMixin", "citadel",